refactor: renamed files to conform to dart styling
