+++ lib/pages/scout/DesktopScout.dart
@@ -2,9 +2,9 @@ import 'package:ax_dapp/dialogs/buy/BuyDialog.dart';
 import 'package:ax_dapp/dialogs/buy/bloc/BuyDialogBloc.dart';
 import 'package:ax_dapp/pages/athlete/bloc/AthletePageBloc.dart';
 import 'package:ax_dapp/pages/scout/Widget%20Factories/AthleteDetailsWidget.dart';
-import 'package:ax_dapp/pages/scout/components/FilterMenuerror.dart';
-import 'package:ax_dapp/pages/scout/components/ScoutLoading.dart';
-import 'package:ax_dapp/pages/scout/components/ScoutLoadingError.dart';
+import 'package:ax_dapp/pages/scout/components/filter_menu_error.dart';
+import 'package:ax_dapp/pages/scout/components/scout_loading.dart';
+import 'package:ax_dapp/pages/scout/components/scout_loading_error.dart';
 import 'package:ax_dapp/pages/scout/dialogs/misc.dart';
 import 'package:ax_dapp/repositories/MlbRepo.dart';
 import 'package:ax_dapp/repositories/subgraph/usecases/GetBuyInfoUseCase.dart';
