+++ lib/pages/scout/components/filter_menu_error.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
 class FilterMenuError extends StatelessWidget {
