+++ lib/pages/scout/components/scout_loading.dart
