+++ lib/pages/scout/components/scout_loading_error.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
 class ScoutLoadingError extends StatelessWidget {
